pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orchesis"
version = "0.6.0"
description = "Agent Runtime Governance Layer — deterministic policy enforcement for AI agent tool calls"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Orchesis Team"},
]
keywords = ["ai", "agents", "security", "mcp", "governance", "policy", "enforcement"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries",
]
dependencies = [
  "PyYAML",
  "click",
  "fastapi",
  "uvicorn",
  "httpx",
  "mcp",
  "cryptography",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mcp-server-sqlite",
  "hypothesis",
  "build",
  "twine",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/orchesis"
Documentation = "https://github.com/YOUR_USERNAME/orchesis/tree/main/docs"
Issues = "https://github.com/YOUR_USERNAME/orchesis/issues"

[project.scripts]
orchesis = "orchesis.cli:main"
orchesis-agent = "orchesis.agent.runner:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
